requirements.txt
smartapi-python
python-dotenv
schedule